mem2reg pass works properly and llc can generate asm from llvm ir properly

+++ java/backend/pass/PassCreator.java
@@ -35,7 +35,7 @@ public class PassCreator
 	}
 
 	public static void createStandardModulePasses(PassManager pm,
-			byte optimizationLevel,
+			int optimizationLevel,
 			boolean optimizeSize,
 			boolean unrollLoops,
 			Pass inliningPass)
